Migrate some pack endpoints
