+++ server/service/endpoint_packs.go
@@ -83,75 +83,6 @@ func makeListPacksEndpoint(svc fleet.Service) endpoint.Endpoint {
 }
 
 ////////////////////////////////////////////////////////////////////////////////
-// Create Pack
-////////////////////////////////////////////////////////////////////////////////
-
-type createPackRequest struct {
-	payload fleet.PackPayload
-}
-
-type createPackResponse struct {
-	Pack packResponse `json:"pack,omitempty"`
-	Err  error        `json:"error,omitempty"`
-}
-
-func (r createPackResponse) error() error { return r.Err }
-
-func makeCreatePackEndpoint(svc fleet.Service) endpoint.Endpoint {
-	return func(ctx context.Context, request interface{}) (interface{}, error) {
-		req := request.(createPackRequest)
-		pack, err := svc.NewPack(ctx, req.payload)
-		if err != nil {
-			return createPackResponse{Err: err}, nil
-		}
-
-		resp, err := packResponseForPack(ctx, svc, *pack)
-		if err != nil {
-			return createPackResponse{Err: err}, nil
-		}
-
-		return createPackResponse{
-			Pack: *resp,
-		}, nil
-	}
-}
-
-////////////////////////////////////////////////////////////////////////////////
-// Modify Pack
-////////////////////////////////////////////////////////////////////////////////
-
-type modifyPackRequest struct {
-	ID      uint
-	payload fleet.PackPayload
-}
-
-type modifyPackResponse struct {
-	Pack packResponse `json:"pack,omitempty"`
-	Err  error        `json:"error,omitempty"`
-}
-
-func (r modifyPackResponse) error() error { return r.Err }
-
-func makeModifyPackEndpoint(svc fleet.Service) endpoint.Endpoint {
-	return func(ctx context.Context, request interface{}) (interface{}, error) {
-		req := request.(modifyPackRequest)
-		pack, err := svc.ModifyPack(ctx, req.ID, req.payload)
-		if err != nil {
-			return modifyPackResponse{Err: err}, nil
-		}
-
-		resp, err := packResponseForPack(ctx, svc, *pack)
-		if err != nil {
-			return modifyPackResponse{Err: err}, nil
-		}
-
-		return modifyPackResponse{
-			Pack: *resp,
-		}, nil
-	}
-}
-
-////////////////////////////////////////////////////////////////////////////////
 // Delete Pack
 ////////////////////////////////////////////////////////////////////////////////
