+++ server/service/handler_test.go
@@ -118,14 +118,6 @@ func TestAPIRoutes(t *testing.T) {
 			uri:  "/api/v1/fleet/packs",
 		},
 		{
-			verb: "POST",
-			uri:  "/api/v1/fleet/packs",
-		},
-		{
-			verb: "PATCH",
-			uri:  "/api/v1/fleet/packs/1",
-		},
-		{
 			verb: "DELETE",
 			uri:  "/api/v1/fleet/packs/1",
 		},
