+++ server/service/service_packs.go
@@ -76,113 +76,6 @@ func (svc *Service) ListPacks(ctx context.Context, opt fleet.PackListOptions) ([
 	return svc.ds.ListPacks(ctx, opt)
 }
 
-func (svc *Service) NewPack(ctx context.Context, p fleet.PackPayload) (*fleet.Pack, error) {
-	if err := svc.authz.Authorize(ctx, &fleet.Pack{}, fleet.ActionWrite); err != nil {
-		return nil, err
-	}
-
-	var pack fleet.Pack
-
-	if p.Name != nil {
-		pack.Name = *p.Name
-	}
-
-	if p.Description != nil {
-		pack.Description = *p.Description
-	}
-
-	if p.Platform != nil {
-		pack.Platform = *p.Platform
-	}
-
-	if p.Disabled != nil {
-		pack.Disabled = *p.Disabled
-	}
-
-	if p.HostIDs != nil {
-		pack.HostIDs = *p.HostIDs
-	}
-
-	if p.LabelIDs != nil {
-		pack.LabelIDs = *p.LabelIDs
-	}
-
-	if p.TeamIDs != nil {
-		pack.TeamIDs = *p.TeamIDs
-	}
-
-	_, err := svc.ds.NewPack(ctx, &pack)
-	if err != nil {
-		return nil, err
-	}
-
-	if err := svc.ds.NewActivity(
-		ctx,
-		authz.UserFromContext(ctx),
-		fleet.ActivityTypeCreatedPack,
-		&map[string]interface{}{"pack_id": pack.ID, "pack_name": pack.Name},
-	); err != nil {
-		return nil, err
-	}
-
-	return &pack, nil
-}
-
-func (svc *Service) ModifyPack(ctx context.Context, id uint, p fleet.PackPayload) (*fleet.Pack, error) {
-	if err := svc.authz.Authorize(ctx, &fleet.Pack{}, fleet.ActionWrite); err != nil {
-		return nil, err
-	}
-
-	pack, err := svc.ds.Pack(ctx, id)
-	if err != nil {
-		return nil, err
-	}
-
-	if p.Name != nil && pack.EditablePackType() {
-		pack.Name = *p.Name
-	}
-
-	if p.Description != nil && pack.EditablePackType() {
-		pack.Description = *p.Description
-	}
-
-	if p.Platform != nil {
-		pack.Platform = *p.Platform
-	}
-
-	if p.Disabled != nil {
-		pack.Disabled = *p.Disabled
-	}
-
-	if p.HostIDs != nil && pack.EditablePackType() {
-		pack.HostIDs = *p.HostIDs
-	}
-
-	if p.LabelIDs != nil && pack.EditablePackType() {
-		pack.LabelIDs = *p.LabelIDs
-	}
-
-	if p.TeamIDs != nil && pack.EditablePackType() {
-		pack.TeamIDs = *p.TeamIDs
-	}
-
-	err = svc.ds.SavePack(ctx, pack)
-	if err != nil {
-		return nil, err
-	}
-
-	if err := svc.ds.NewActivity(
-		ctx,
-		authz.UserFromContext(ctx),
-		fleet.ActivityTypeEditedPack,
-		&map[string]interface{}{"pack_id": pack.ID, "pack_name": pack.Name},
-	); err != nil {
-		return nil, err
-	}
-
-	return pack, err
-}
-
 func (svc *Service) DeletePack(ctx context.Context, name string) error {
 	if err := svc.authz.Authorize(ctx, &fleet.Pack{}, fleet.ActionWrite); err != nil {
 		return err
