+++ server/service/transport_packs.go
@@ -6,28 +6,6 @@ import (
 	"net/http"
 )
 
-func decodeCreatePackRequest(ctx context.Context, r *http.Request) (interface{}, error) {
-	var req createPackRequest
-	if err := json.NewDecoder(r.Body).Decode(&req.payload); err != nil {
-		return nil, err
-	}
-
-	return req, nil
-}
-
-func decodeModifyPackRequest(ctx context.Context, r *http.Request) (interface{}, error) {
-	id, err := uintFromRequest(r, "id")
-	if err != nil {
-		return nil, err
-	}
-	var req modifyPackRequest
-	if err := json.NewDecoder(r.Body).Decode(&req.payload); err != nil {
-		return nil, err
-	}
-	req.ID = uint(id)
-	return req, nil
-}
-
 func decodeDeletePackRequest(ctx context.Context, r *http.Request) (interface{}, error) {
 	name, err := stringFromRequest(r, "name")
 	if err != nil {
